Move random real to fixed pt specific class
