+++ core/src/main/java/dk/alexandra/fresco/lib/real/fixed/AdvancedFixedNumeric.java
@@ -1,6 +1,7 @@
 package dk.alexandra.fresco.lib.real.fixed;
 
 import dk.alexandra.fresco.framework.DRes;
+import dk.alexandra.fresco.framework.builder.numeric.AdvancedNumeric.RandomAdditiveMask;
 import dk.alexandra.fresco.framework.builder.numeric.ProtocolBuilderNumeric;
 import dk.alexandra.fresco.framework.value.SInt;
 import dk.alexandra.fresco.lib.real.DefaultAdvancedRealNumeric;
@@ -8,6 +9,8 @@ import dk.alexandra.fresco.lib.real.SReal;
 import java.math.BigDecimal;
 
 public class AdvancedFixedNumeric extends DefaultAdvancedRealNumeric {
+
+  private static final int DEFAULT_RANDOM_BITS = 64;
 
   public AdvancedFixedNumeric(ProtocolBuilderNumeric builder) {
     super(builder);
@@ -29,5 +32,15 @@ public class AdvancedFixedNumeric extends DefaultAdvancedRealNumeric {
       return result;
     });
 
+  }
+
+  @Override
+  public DRes<SReal> random() {
+    return builder.seq(seq -> {
+      DRes<RandomAdditiveMask> random = seq.advancedNumeric().additiveMask(DEFAULT_RANDOM_BITS);
+      return random;
+    }).seq((seq, random) -> {
+      return () -> new SFixed(random.random, DEFAULT_RANDOM_BITS);
+    });
   }
 }
